add entity User
