+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/tatsuo48/go_ddd
 
 go 1.15
+
+require github.com/google/uuid v1.1.2
