+++ main.go
@@ -1,30 +1,13 @@
 package main
 
 import (
-	"errors"
 	"fmt"
-	"unicode/utf8"
 )
-
-type CircleID string
-
-type CircleName string
-
-func NewCircleName(s string) (CircleName, error) {
-	if utf8.RuneCountInString(s) < 3 {
-		return CircleName(""), errors.New("サークル名は3文字以上です。")
-	} else if utf8.RuneCountInString(s) > 20 {
-		return CircleName(""), errors.New("サークル名は20文字以下です。")
-	}
-	return CircleName(s), nil
-}
-
-func (c CircleName) Equals(otherCircleName CircleName) bool {
-	return string(c) == string(otherCircleName)
-}
 
 func main() {
 	a, _ := NewCircleName("test")
 	b, _ := NewCircleName("test")
 	fmt.Println(a.Equals(b))
+	u, _ := NewUser("test")
+	fmt.Println(u)
 }
